Clients can now time out.
Added
- Clients that don't send the proper CQ Online message
  for more than 118 seconds, will be considered offline

Changed
- Changed the way offline and online calls are stored and handled

The new way:
Every Client is stored in a list, when the table refreshes, it'll sort
them into online calls and offline calls and display them. Each time
the own client sents out an online CQ call, it will check for clients
which didn't send out the messages for over 118 seconds (almost 2 minutes).
If the client find any clients it will mark them as offline and reload the
online table...